refactor font rendering more

+++ src/main/java/de/bixilon/minosoft/gui/rendering/font/renderer/element/TextRenderInfo.kt
@@ -21,7 +21,7 @@ import de.bixilon.minosoft.util.logging.LogMessageType
 
 class TextRenderInfo(
     val fontAlignment: HorizontalAlignments,
-    val charHeight: Int,
+    val charHeight: Float,
     val charMargin: Int,
     val scale: Float,
     val shadow: Boolean,
